% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prauc-package.R
\docType{package}
\name{prauc-package}
\alias{prauc}
\alias{prauc-package}
\title{prauc: PR AUC Estimates With Confidence Interval}
\description{
Compute average precision-recall with logit confidence interval, following Boyd et al. (2013). Designed to integrate with dplyr syntax.
}
\references{
Boyd, K., Eng, K. H., & Page, C. D. (2013).
Area under the precision-recall curve: point estimates and confidence intervals.
In Joint European conference on machine learning and knowledge discovery in databases (pp. 451-466).
Springer, Berlin, Heidelberg.
}
\author{
\strong{Maintainer}: Louis Duchemin \email{ls.duchemin@gmail.com} (\href{https://orcid.org/0000-0002-7984-9554}{ORCID})

Authors:
\itemize{
  \item Philippe Veber
}

}
\keyword{internal}
