% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prauc.R
\name{prauc_avg}
\alias{prauc_avg}
\title{Compute average PR AUC}
\usage{
prauc_avg(scores, oracle, pval = T)
}
\arguments{
\item{scores}{vector of scores}

\item{oracle}{vector of labels T/F}

\item{pval}{if ture, transform scores as ( 1 - score )}
}
\value{
numeric value of average PR AUC
}
\description{
Compute average PR AUC
}
