% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prauc.R
\name{prauc_logit_ci}
\alias{prauc_logit_ci}
\title{Compute logit confidence interval}
\usage{
prauc_logit_ci(avg_auc, alpha, n_pos)
}
\arguments{
\item{avg_auc}{average PR AUC estimate}

\item{alpha}{risk for type 1 error}

\item{n_pos}{number of true positive labels}
}
\value{
a named 2-vector of confidence interval bounds
}
\description{
Compute logit confidence interval
}
